feat 南京接口 应急采购 目录 订单 明细

+++ tripartite-his/tripartite-nanjing-procurement/tripartite-nanjing-procurement-openapi/src/main/java/work/gaigeshen/tripartite/nanjing/procurement/openapi/parameters/province/NanJingProcurementEmergencyPlanCreateInputData.java
@@ -73,7 +73,7 @@ public class NanJingProcurementEmergencyPlanCreateInputData implements NanJingPr
         /**
          * 采购数量（必须 > 0）
          */
-        private Integer purcCnt;
+        private BigDecimal purcCnt;
 
         /**
          * 医院采购价
